+++ NEWS.md
@@ -3,6 +3,9 @@
 * `data_arrange()` now works with dataframes that were grouped using `data_group()`
   (#274).
 
+* `data_read()` gains a `convert_factors` argument, to turn off automatic
+  conversion from numeric variables into factors.
+
 # datawizard 0.6.1
 
 * Updates tests for upcoming changes in the `{tidyselect}` package (#267).