+++ NEWS.md
@@ -1,5 +1,8 @@
 # datawizard (development version)
 
+* `data_read()` gains a `convert_factors` argument, to turn off automatic
+  conversion from numeric variables into factors.
+
 # datawizard 0.6.1
 
 * Updates tests for upcoming changes in the `{tidyselect}` package (#267).